LF-18506-4 net: enetc: fix incorrect offset of hash_tbl in VLAN hash filter message
The hash_tbl field in struct enetc_msg_vlan_hash_filter is expected to
start at offset 0x14 according to the message format specification.
However, since hash_tbl was previously defined as u64[], the compiler
inserts implicit padding after the existing u8 resv[3] field to satisfy
the 8-byte alignment requirement of u64, causing hash_tbl to start at
offset 0x18 instead.

Fix this by replacing u64 hash_tbl[] with u32 hash_tbl[], so that
hash_tbl follows directly after the resv[3] field with no implicit
padding, correctly placing it at offset 0x14.

Fixes: 9999ce0a3040 ("net: enetc: add VLAN filtering support for VFs")

+++ drivers/net/ethernet/freescale/enetc/enetc_mailbox.h
@@ -199,7 +199,7 @@ struct enetc_msg_vlan_hash_filter {
 	struct enetc_msg_header hdr;
 	u8 size;
 	u8 resv[3];
-	u64 hash_tbl[];
+	u32 hash_tbl[];
 };
 
 /* message format of class_id 0x21, cmd_id: 0x4,
